fix: resolve sintaxis error in calendar event attende service
